Make feedback banner use the configuration in the locale file
The feedback banner has a locale configuration value, but we had stopped using it at some point in favour of a variable in the phase banner template file.

This came as a bit of a surprise, resulting in the feedback banner being live unexpectedly.
