+++ test/integration/whats_new_test.rb
@@ -19,15 +19,4 @@ class WhatsNewTest < ActionDispatch::IntegrationTest
       end
     end
   end
-
-  test "uses the show_banner value in the whats_new.yml file to determine whether to render the banner" do
-    login_as create(:gds_editor)
-    get admin_whats_new_path
-
-    if I18n.t("admin.whats_new.show_banner") || I18n.t("admin.feedback.show_banner")
-      assert_select "#whats_new_banner.gem-c-phase-banner", count: 1
-    else
-      assert_select "#whats_new_banner.gem-c-phase-banner", count: 0
-    end
-  end
 end
